WIP: possible race in sidejob fix
Beware, this is unbenched and messy code, this commit is just to save
WIP. The change here is to serialise access to each resource worker's
ETS table through a gen_server. There were certain sets of actions (see
`sidejob_worker:is_available`) that should be atomic but could
interleave. However, maybe the original author made the trade off for a
loosy-goosy limit for greater performance. We may never know. There is
still a failing property to do with `which_children`.